working on timing data

+++ io.hh
@@ -8,7 +8,7 @@
 std::vector<int> genInputNumbers(int length, int max);
 
 unsigned long getCurrentTime();
-void outputTable(std::vector<std::vector<double>> timingData, std::vector<std::string> colLabels, std::vector<std::string> rowLabels);
+void outputTable(std::vector<std::vector<int>> timingData, std::vector<std::string> colLabels, std::vector<std::string> rowLabels);
 void outputColLabels(std::vector<std::string> colLabels);
 void outputDivider(int size);
 void outputRowLabel(std::string label);
